refactor and move dashboard action to hbxprofile controller
