+++ spec/domain/operations/hbx_admin/dry_run/individual/analyzer_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe ::Operations::HbxAdmin::DryRun::IndividualMarketBenefits do
+RSpec.describe ::Operations::HbxAdmin::DryRun::Individual::Analyzer do
   let!(:hbx_profile) {FactoryBot.create(:hbx_profile)}
   let!(:benefit_sponsorship) { FactoryBot.create(:benefit_sponsorship, hbx_profile: hbx_profile) }
   let!(:benefit_coverage_period_previous_year) do
@@ -27,9 +27,29 @@ RSpec.describe ::Operations::HbxAdmin::DryRun::IndividualMarketBenefits do
                      open_enrollment_end_on: ((TimeKeeper.date_of_record + 1.year).beginning_of_year + 1.month))
   end
 
+  let!(:application) do
+    FactoryBot.create(:financial_assistance_application, hbx_id: '200000126', aasm_state: "determined", family_id: "family.id", assistance_year: benefit_coverage_period_next_year.start_on.year, predecessor_id: "predecessor_id")
+  end
+  let!(:applicant) do
+    FactoryBot.create(:financial_assistance_applicant,
+                      eligibility_determination_id: nil,
+                      person_hbx_id: '1629165429385938',
+                      is_primary_applicant: true,
+                      first_name: 'esi',
+                      last_name: 'evidence',
+                      ssn: "518124854",
+                      dob: Date.new(1988, 11, 11),
+                      family_member_id: "family.primary_family_member.id",
+                      application: application)
+  end
+
+  before do
+    benefit_sponsorship.benefit_coverage_periods = []
+    benefit_sponsorship.benefit_coverage_periods = [benefit_coverage_period_previous_year, benefit_coverage_period_this_year, benefit_coverage_period_next_year]
+  end
 
   context 'with valid data' do
-    it 'should return renewal benefit coverage params' do
+    it 'should return a skeleton hash for application states by coverage years' do
       result = described_class.new.call
       expect(result).to be_success
     end
